[UPDATED] Split bindings to renderer
