+++ engine/bindings/private/audio/audio_bindings.hpp
@@ -1,4 +1,5 @@
 #pragma once
-#include <wren_common.hpp>
+
+#include "wren_common.hpp"
 
 void BindAudioAPI(wren::ForeignModule& module);
