+++ engine/bindings/private/renderer/animation_bindings.hpp
@@ -1,0 +1,5 @@
+#pragma once
+
+#include "wren_common.hpp"
+
+void BindAnimationAPI(wren::ForeignModule& module);
